syn: simply send

+++ include/functions.h
@@ -8,7 +8,7 @@ t_addr get_interface(sa_family_t family);
 t_addrinfo get_info(char *host);
 
 // checksum.c
-unsigned short tcp_checksum(t_packet packet, size_t value);
+unsigned short tcp_checksum(t_packet packet, unsigned short value);
 
 // tcp.c
 void send_packet_tcp(unsigned short port);
